+++ lib/compat/wordpress-6.0/block-patterns.php
@@ -15,7 +15,7 @@ if ( ! function_exists( '_register_remote_theme_patterns' ) ) {
 			return;
 		}
 
-		if ( ! WP_Theme_JSON_Resolver_Gutenberg::theme_has_support() ) {
+		if ( ! wp_theme_has_theme_json() ) {
 			return;
 		}