+++ packages/block-editor/src/store/reducer.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { reduce, omit, mapValues, isEqual, isEmpty, omitBy } from 'lodash';
+import { omit, mapValues, isEqual, isEmpty } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -417,15 +417,15 @@ const withBlockTree =
 				break;
 			}
 			case 'SAVE_REUSABLE_BLOCK_SUCCESS': {
-				const updatedBlockUids = Object.keys(
-					omitBy( newState.attributes, ( attributes, clientId ) => {
+				const updatedBlockUids = Object.entries( newState.attributes )
+					.filter( ( [ clientId, attributes ] ) => {
 						return (
-							newState.byClientId[ clientId ].name !==
-								'core/block' ||
-							attributes.ref !== action.updatedId
+							newState.byClientId[ clientId ].name ===
+								'core/block' &&
+							attributes.ref === action.updatedId
 						);
 					} )
-				);
+					.map( ( [ clientId ] ) => clientId );
 
 				newState.tree = updateParentInnerBlocksInTree(
 					newState,
@@ -683,30 +683,22 @@ const withReplaceInnerBlocks = ( reducer ) => ( state, action ) => {
 		// will be deleted entirely from its entity.
 		stateAfterInsert.order = {
 			...stateAfterInsert.order,
-			...reduce(
-				nestedControllers,
-				( result, value, key ) => {
-					if ( state.order[ key ] ) {
-						result[ key ] = state.order[ key ];
-					}
-					return result;
-				},
-				{}
-			),
+			...Object.keys( nestedControllers ).reduce( ( result, key ) => {
+				if ( state.order[ key ] ) {
+					result[ key ] = state.order[ key ];
+				}
+				return result;
+			}, {} ),
 		};
 		stateAfterInsert.tree = {
 			...stateAfterInsert.tree,
-			...reduce(
-				nestedControllers,
-				( result, value, _key ) => {
-					const key = `controlled||${ _key }`;
-					if ( state.tree[ key ] ) {
-						result[ key ] = state.tree[ key ];
-					}
-					return result;
-				},
-				{}
-			),
+			...Object.keys( nestedControllers ).reduce( ( result, _key ) => {
+				const key = `controlled||${ _key }`;
+				if ( state.tree[ key ] ) {
+					result[ key ] = state.tree[ key ];
+				}
+				return result;
+			}, {} ),
 		};
 	}
 	return stateAfterInsert;
@@ -873,24 +865,22 @@ export const blocks = pipe(
 				const next = action.clientIds.reduce(
 					( accumulator, id ) => ( {
 						...accumulator,
-						[ id ]: reduce(
+						[ id ]: Object.entries(
 							action.uniqueByBlock
 								? action.attributes[ id ]
-								: action.attributes,
-							( result, value, key ) => {
-								// Consider as updates only changed values.
-								if ( value !== result[ key ] ) {
-									result = getMutateSafeObject(
-										state[ id ],
-										result
-									);
-									result[ key ] = value;
-								}
+								: action.attributes ?? {}
+						).reduce( ( result, [ key, value ] ) => {
+							// Consider as updates only changed values.
+							if ( value !== result[ key ] ) {
+								result = getMutateSafeObject(
+									state[ id ],
+									result
+								);
+								result[ key ] = value;
+							}
 
-								return result;
-							},
-							state[ id ]
-						),
+							return result;
+						}, state[ id ] ),
 					} ),
 					{}
 				);
@@ -1056,8 +1046,7 @@ export const blocks = pipe(
 					} ),
 					( nextState ) =>
 						mapValues( nextState, ( subState ) =>
-							reduce(
-								subState,
+							Object.values( subState ).reduce(
 								( result, clientId ) => {
 									if ( clientId === clientIds[ 0 ] ) {
 										return [
@@ -1159,6 +1148,26 @@ export const blocks = pipe(
 		return state;
 	},
 } );
+
+/**
+ * Reducer returning visibility status of block interface.
+ *
+ * @param {boolean} state  Current state.
+ * @param {Object}  action Dispatched action.
+ *
+ * @return {boolean} Updated state.
+ */
+export function isBlockInterfaceHidden( state = false, action ) {
+	switch ( action.type ) {
+		case 'HIDE_BLOCK_INTERFACE':
+			return true;
+
+		case 'SHOW_BLOCK_INTERFACE':
+			return false;
+	}
+
+	return state;
+}
 
 /**
  * Reducer returning typing state.
@@ -1809,6 +1818,7 @@ export function temporarilyEditingAsBlocks( state = '', action ) {
 export default combineReducers( {
 	blocks,
 	isTyping,
+	isBlockInterfaceHidden,
 	draggedBlocks,
 	selection,
 	isMultiSelecting,