+++ packages/components/src/navigator/navigator-screen/component.tsx
@@ -4,12 +4,13 @@
 import type { ForwardedRef } from 'react';
 // eslint-disable-next-line no-restricted-imports
 import { motion, MotionProps } from 'framer-motion';
+import { css } from '@emotion/react';
 
 /**
  * WordPress dependencies
  */
 import { focus } from '@wordpress/dom';
-import { useContext, useEffect, useRef } from '@wordpress/element';
+import { useContext, useEffect, useMemo, useRef } from '@wordpress/element';
 import {
 	useReducedMotion,
 	useMergeRefs,
@@ -26,6 +27,7 @@ import {
 	useContextSystem,
 	WordPressComponentProps,
 } from '../../ui/context';
+import { useCx } from '../../utils/hooks/use-cx';
 import { View } from '../../view';
 import { NavigatorContext } from '../context';
 import type { NavigatorScreenProps } from '../types';
@@ -42,7 +44,10 @@ type Props = Omit<
 	keyof MotionProps
 >;
 
-function NavigatorScreen( props: Props, forwardedRef: ForwardedRef< any > ) {
+function UnconnectedNavigatorScreen(
+	props: Props,
+	forwardedRef: ForwardedRef< any >
+) {
 	const { children, className, path, ...otherProps } = useContextSystem(
 		props,
 		'NavigatorScreen'
@@ -54,6 +59,21 @@ function NavigatorScreen( props: Props, forwardedRef: ForwardedRef< any > ) {
 	const wrapperRef = useRef< HTMLDivElement >( null );
 
 	const previousLocation = usePrevious( location );
+
+	const cx = useCx();
+	const classes = useMemo(
+		() =>
+			cx(
+				css( {
+					// Ensures horizontal overflow is visually accessible.
+					overflowX: 'auto',
+					// In case the root has a height, it should not be exceeded.
+					maxHeight: '100%',
+				} ),
+				className
+			),
+		[ className, cx ]
+	);
 
 	// Focus restoration
 	const isInitialLocation = location.isInitial && ! location.isBack;
@@ -119,7 +139,7 @@ function NavigatorScreen( props: Props, forwardedRef: ForwardedRef< any > ) {
 		return (
 			<View
 				ref={ mergedWrapperRef }
-				className={ className }
+				className={ classes }
 				{ ...otherProps }
 			>
 				{ children }
@@ -165,7 +185,7 @@ function NavigatorScreen( props: Props, forwardedRef: ForwardedRef< any > ) {
 	return (
 		<motion.div
 			ref={ mergedWrapperRef }
-			className={ className }
+			className={ classes }
 			{ ...otherProps }
 			{ ...animatedProps }
 		>
@@ -208,9 +228,9 @@ function NavigatorScreen( props: Props, forwardedRef: ForwardedRef< any > ) {
  * );
  * ```
  */
-const ConnectedNavigatorScreen = contextConnect(
-	NavigatorScreen,
+export const NavigatorScreen = contextConnect(
+	UnconnectedNavigatorScreen,
 	'NavigatorScreen'
 );
 
-export default ConnectedNavigatorScreen;
+export default NavigatorScreen;