+++ packages/react-native-editor/CHANGELOG.md
@@ -11,8 +11,15 @@ For each user feature we should also add a importance categorization label  to i
 
 ## Unreleased
 
+## 1.85.0
+-   [*] [iOS] Fixed iOS Voice Control support within Image block captions. [#44850]
+
+## 1.84.1
+-   [**] Native inner blocks merge where appropriate [#45048]
+
 ## 1.84.0
 -   [*] Upgrade compile and target sdk version to Android API 31 [#44610]
+-   [*] [iOS] Fixed iOS Voice Control support within Image block captions. [#44850]
 
 ## 1.83.0
 * No User facing changes *